Split tests for fips/nofips variants
